+++ src/main.cpp
@@ -1,10 +1,11 @@
 #include "game.h"
-#include "enableansi.h"
+#include "consolesettings.h"
 
 int main()
 {
     #ifdef _WIN32
         enableANSI();
+        enableUTF8();
     #endif
     
     Game game = Game();